QOLOE-52 move main to js folder qld.bootstrap.js
